analysis-tool/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pywebview>=5.0
requests>=2.31.0
python-multipart>=0.0.9
